Fix build compatibility issues for internal FFmpeg
Add conditional compilation guards for FFmpeg profile constants and
codec IDs that are not present in the internal FFmpeg but are added
via patches when using the depends-built FFmpeg:

- Guard FF_PROFILE_DTS_HD_MA_X and FF_PROFILE_DTS_HD_MA_X_IMAX
- Guard FF_PROFILE_EAC3_DDP_ATMOS and FF_PROFILE_TRUEHD_ATMOS
- Guard AV_CODEC_ID_H264_MVC for MVC 3D video support
- Guard HAS_LIBAMCODEC-specific aml_set_audio_passthrough call
- Guard HAVE_LIBBLURAY-specific CDVDInputStreamBluray usage
- Fix std::min type mismatches in JSONServiceDescription and GUITextBox

These changes allow building with internal FFmpeg which lacks
the custom Kodi patches for DTS-X, Dolby Atmos, and H.264 MVC
detection.

+++ xbmc/interfaces/json-rpc/JSONServiceDescription.cpp
@@ -796,7 +796,7 @@ JSONRPC_STATUS JSONSchemaTypeDefinition::Check(const CVariant& value,
       // are either no more schemas in the "items"
       // array or no more elements in the value's array
       unsigned int arrayIndex;
-      for (arrayIndex = 0; arrayIndex < std::min(items.size(), value.size()); arrayIndex++)
+      for (arrayIndex = 0; arrayIndex < std::min(items.size(), static_cast<size_t>(value.size())); arrayIndex++)
       {
         JSONRPC_STATUS status = items.at(arrayIndex)->Check(value[arrayIndex], outputValue[arrayIndex], errorData["property"]);
         if (status != OK)
